feat: define our overlay outside of the X config
We want to be able to use our nix overlay outside of a shell in the X server.

This change enables us to define the overlay in the NIX_PATH when we
start a shell (bash/zsh) on the machine (and not when we start a X session on the
machine).

This change has been tested on my devbox where I only ssh into.